+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 # Changelog
 
-## next
+## 4.2.0
 
 ### Improvements
 
-- Added a `longName` field to the `Author` type, which returns the display name set by the author at a particular path in a workspace (see https://github.com/earthstar-project/earthstar/blob/master/docs/vocabulary.md#author)
+- Added a `displayName` field to the `Author` type, which returns the display name set by the author at a particular path in a workspace (see https://github.com/earthstar-project/earthstar/blob/master/docs/vocabulary.md#author)
 
 ## 4.1.1